+++ sdk/appcenter-push/src/main/java/com/microsoft/appcenter/push/Push.java
@@ -271,16 +271,27 @@ public class Push extends AbstractAppCenterService {
      */
     @SuppressWarnings("WeakerAccess") /* protected so that Xamarin can use it. */
     protected synchronized void onTokenRefresh(final String pushToken) {
-        if (pushToken != null && !pushToken.equals(mLatestPushToken)) {
-            AppCenterLog.debug(LOG_TAG, "Push token refreshed: " + pushToken);
-            mLatestPushToken = pushToken;
+        Runnable disabledRunnable = new Runnable() {
+
+            @Override
+            public void run() {
+                AppCenterLog.debug(LOG_TAG, "Push token refreshed before start/enabled, ignoring.");
+            }
+        };
+        if (mContext == null) {
+            disabledRunnable.run();
+        } else {
             post(new Runnable() {
 
                 @Override
                 public void run() {
-                    enqueuePushInstallationLog(pushToken);
+                    if (pushToken != null && !pushToken.equals(mLatestPushToken)) {
+                        AppCenterLog.debug(LOG_TAG, "Push token refreshed: " + pushToken);
+                        mLatestPushToken = pushToken;
+                        enqueuePushInstallationLog(pushToken);
+                    }
                 }
-            });
+            }, disabledRunnable, disabledRunnable);
         }
     }